aidit/aidit/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pillow==10.1.0
python-multipart==0.0.6
openai==1.12.0
numpy<2.0.0
opencv-python==4.8.1.78
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.3.0
videodb
httpx==0.25.2

